% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difference_plot.R
\name{difference_plot}
\alias{difference_plot}
\title{Draw a difference plot (aka a Bland-Altman plot)}
\usage{
difference_plot(
  data,
  sample_col,
  result_col,
  analyser_col,
  difference = c("absolute", "relative"),
  comparison = "mean",
  title = "",
  subtitle = "",
  units = NULL
)
}
\arguments{
\item{data}{A data frame, where each row is a measurement of a sample by an
analyser.}

\item{sample_col}{The name of a column in \code{data} that identifies each sample.}

\item{result_col}{The name of a column in \code{data} that identifies the result.}

\item{analyser_col}{The name of a column in \code{data} that identifies the analyser.}

\item{difference}{A string, either "absolute" (default) or "relative".
Whether the difference on the y-axis is plotted in the reported units, or
as a \% of the comparison.}

\item{comparison}{A string, either "mean" (default) or a value from the
\code{analyser_col} column in \code{data}. If "mean", then x-axis will be the mean of
all results for each sample. If it is an analyser, then the x-axis will be
the result for that analyser,and no points will be plotted for that
analyser (because the difference will always be zero).}

\item{title}{A string.}

\item{subtitle}{A string.}

\item{units}{A string.}
}
\value{
A ggplot2 object
}
\description{
Draw a difference plot (aka a Bland-Altman plot)
}
\examples{
## Plot the absolute difference (default) against the mean result (default)
difference_plot(hidx_eqa,
                sample_col = sample,
                result_col = h_index,
                analyser_col = analyser)

## Plot the relative difference against the EQA method mean (which we've
## recorded as if it were an analyser)
difference_plot(hidx_eqa,
                sample_col = sample,
                result_col = h_index,
                analyser_col = analyser,
                comparison = "EQA Method Mean",
                difference = "relative")
}
